+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 - 2021-04-08
+### Added
+- #### BREAKING
+
+    `maxNumberItems` can be used to limit the number of items in a carousel (this used to be implicitly set to 10). If not set you can add unlimited number of items.
+
+
+### Changed
+- addItemAt can now return a bool. Returning true or null indicates that an item was added and that the selected item should be updated. Returning false means that the item ended up not being added, and thus selected item shouldn't be updated.
+
+    Seeing as null is an appropriate return value, this won't be a breaking change.
+
+
 ## 0.2.0 - 2021-03-31
 ### Changed
 - addItemAt can now be used async. Meaning that onItemSelected won't be called until it completes